snomed_diz_llm/requirements.txt
streamlit
langchain==0.3.0
langchain-community==0.3.0
pandas==2.2.2
numpy==1.26.4
python-dotenv
json-repair
transformers==4.44.2
sentence-transformers==3.0.1
neo4j==5.20.0
faiss-cpu==1.7.4
openai>=1.40.0
tiktoken
requests
PyPDF2
torch==2.2.2
pyvis
tqdm
huggingface_hub
llama-index-core
llama-index-llms-openai-like
llama-index

